Rewrite djinn in Go for improved performance

+++ djinn/count/count.xam.go
@@ -10,7 +10,7 @@ import (
 
 func CountXam(infile string, invalid bool, threads int) error {
 	infile = xam.FileOrStdin(infile)
-	set := make(map[string]int16, 7_000_000)
+	set := make(map[string]int, 7_000_000)
 
 	// ── open reader ───────────────────────────────────────────────────────────
 	recChan, _ := xam.NewXamReaderChan(infile, xam.ChanCap, xam.IoBuf, threads)
